Extract route-map into module
